Example Avogadro after creating combine.molden
My script combine.py creates an file that can be read by Avogadro to animate vibrational modes, and molecular orbitals. It still requires you to go in to Build>Carteisian Editor and re-input the molecular geometry, because for some reason that doesn't get loaded correctly. But the MOs and the IR loads well. The IR intensities don't load, but the frequencies do. I don't know how to fix that yet...